refactor(timeline): extract smart-trim hover state into useSmartTrimHover
Move hoveredEdge / smartTrimIntent / smartBodyIntent state machines,
the activeTool-change reset effect, the mouse-move intent resolution,
and handleMouseLeave into a hook scoped to the TimelineItem.

The smartTrimIntentRef is still exposed so pointer handlers
(handleClick, handleMouseDown, handleSmartTrimStart) can read the
latest intent at gesture-start without re-render churn.
